[Misc] Move utils to avoid conflicts with stdlib, and move tests (#27169)
